Reorganize theme.
