+++ package.json
@@ -8,17 +8,17 @@
     "dev": "npm run watch-scss",
     "compile-scss": "sass styles/avant.scss styles/avant.css --style=expanded",
     "watch-scss": "sass --watch styles/avant.scss styles/avant.css --style=expanded",
-    "theme:help": "node scripts/theme-utils.js help",
-    "theme:template": "node scripts/theme-utils.js template",
-    "theme:template:full": "node scripts/theme-utils.js template-full",
-    "theme:validate": "node scripts/theme-utils.js validate",
-    "theme:docs": "node scripts/theme-utils.js docs",
-    "theme:list": "node scripts/theme-utils.js list-vars",
-    "theme:examples": "node scripts/theme-utils.js create-examples",
-    "theme:mappings": "node scripts/theme-utils.js generate-mappings",
+    "theme:help": "node scripts/themes/theme-utils.js help",
+    "theme:template": "node scripts/themes/theme-utils.js template",
+    "theme:template:full": "node scripts/themes/theme-utils.js template-full",
+    "theme:validate": "node scripts/themes/theme-utils.js validate",
+    "theme:docs": "node scripts/themes/theme-utils.js docs",
+    "theme:list": "node scripts/themes/theme-utils.js list-vars",
+    "theme:examples": "node scripts/themes/theme-utils.js create-examples",
+    "theme:mappings": "node scripts/themes/theme-utils.js generate-mappings",
     "docs:themes": "npm run theme:docs > styles/themes/THEME-VARIABLES.md",
     "docs:update": "npm run docs:themes",
-    "validate:theme": "node scripts/theme-utils.js validate",
+    "validate:theme": "node scripts/themes/theme-utils.js validate",
     "test:themes": "npm run theme:examples && npm run theme:validate styles/themes/examples/minimal-example.json && npm run theme:validate styles/themes/examples/complete-example.json",
     "clean": "rm -f styles/avant.css.map",
     "prepare": "npm run build"
