+++ scripts/avant.js
@@ -3,7 +3,7 @@
 console.log("Avant | Loading game system...");
 
 // Import theme manager
-import { AvantThemeManager } from './theme-manager.js';
+import { AvantThemeManager } from './themes/theme-manager.js';
 
 /**
  * Actor Data Model for Avant - v12/v13 Compatible
