+++ scripts/themes/theme-config.js
