+++ scripts/themes/theme-manager.js
